Refactor tsv output script, call directly from code rather than shelling out
